+++ handlers/tenantServiceV2MonitoredObjectsAPIHandler.go
@@ -45,6 +45,31 @@ func HandleGetAllMonitoredObjectsV2(allowedRoles []string, tenantDB datastore.Te
 	}
 }
 
+// HandleGetMonitoredObjectListV2 - retrieve all monitored object ids for a Tenant based on the specified search criteria
+func HandleGetFilteredMonitoredObjectListV2(allowedRoles []string, tenantDB datastore.TenantServiceDatastore) func(params tenant_provisioning_service_v2.GetFilteredMonitoredObjectListV2Params) middleware.Responder {
+	return func(params tenant_provisioning_service_v2.GetFilteredMonitoredObjectListV2Params) middleware.Responder {
+		// Do the work
+		startTime, responseCode, response, err := doGetFilteredMonitoredObjectListV2(allowedRoles, tenantDB, params)
+
+		// Success Response
+		if responseCode == http.StatusOK {
+			return tenant_provisioning_service_v2.NewGetFilteredMonitoredObjectListV2OK().WithPayload(response)
+		}
+
+		// Error Responses
+		errorMessage := reportAPIError(err.Error(), startTime, responseCode, mon.GetAllMonObjStr, mon.APICompleted, mon.TenantAPICompleted)
+		switch responseCode {
+		case http.StatusForbidden:
+			return tenant_provisioning_service_v2.NewGetAllMonitoredObjectsV2Forbidden().WithPayload(errorMessage)
+		case http.StatusNotFound:
+			return tenant_provisioning_service_v2.NewGetAllMonitoredObjectsV2NotFound().WithPayload(errorMessage)
+		default:
+			return tenant_provisioning_service_v2.NewGetAllMonitoredObjectsV2InternalServerError().WithPayload(errorMessage)
+
+		}
+	}
+}
+
 // HandleCreateMonitoredObjectV2 - create a new Monitored Object
 func HandleCreateMonitoredObjectV2(allowedRoles []string, tenantDB datastore.TenantServiceDatastore) func(params tenant_provisioning_service_v2.CreateMonitoredObjectV2Params) middleware.Responder {
 	return func(params tenant_provisioning_service_v2.CreateMonitoredObjectV2Params) middleware.Responder {
@@ -438,6 +463,65 @@ func doGetAllMonitoredObjectsV2(allowedRoles []string, tenantDB datastore.Tenant
 
 	reportAPICompletionState(startTime, http.StatusOK, mon.GetAllMonObjStr, mon.APICompleted, mon.TenantAPICompleted)
 	logger.Log.Infof("Retrieved %d %ss", len(converted.Data), tenmod.TenantMonitoredObjectStr)
+	return startTime, http.StatusOK, &converted, nil
+}
+
+func doGetFilteredMonitoredObjectListV2(allowedRoles []string, tenantDB datastore.TenantServiceDatastore, params tenant_provisioning_service_v2.GetFilteredMonitoredObjectListV2Params) (time.Time, int, *swagmodels.MonitoredObjectFilteredListResponse, error) {
+	tenantID := params.HTTPRequest.Header.Get(XFwdTenantId)
+	isAuthorized, startTime := authorizeRequest(fmt.Sprintf("Fetching all %s: %s", tenmod.TenantMonitoredObjectKeysStr, tenantID), params.HTTPRequest, allowedRoles, mon.APIRecieved, mon.TenantAPIRecieved)
+
+	if !isAuthorized {
+		err := fmt.Errorf("Get all %s operation not authorized for role: %s", tenmod.TenantMonitoredObjectKeysStr, params.HTTPRequest.Header.Get(XFwdUserRoles))
+		return startTime, http.StatusForbidden, nil, err
+	}
+
+	if tenantID == "" {
+		return startTime, http.StatusBadRequest, nil, fmt.Errorf("Invalid request missing tenant ID")
+	}
+
+	meta := make(map[string][]string)
+
+	requestBytes, err := json.Marshal(params.Body.Meta)
+	if err != nil || params.Body == nil {
+		return startTime, http.StatusBadRequest, nil, fmt.Errorf("Invalid request body: %s", models.AsJSONString(params.Body))
+	}
+	// Convert the request JSON into a map
+	err = json.Unmarshal(requestBytes, &meta)
+	if err != nil {
+		errResp := fmt.Errorf("Unable to unmarshal metadata for %s request for tenant %s: %s", tenmod.TenantMonitoredObjectKeysStr, tenantID, err.Error())
+		return startTime, http.StatusInternalServerError, nil, errResp
+	}
+
+	logger.Log.Debugf("Retrieving %s for tenant %s with meta filters %v", tenmod.TenantMonitoredObjectStr, tenantID, meta)
+
+	resourceIdentifierList, err := tenantDB.GetFilteredMonitoredObjectList(tenantID, meta)
+
+	if err != nil {
+		if checkForNotFound(err.Error()) {
+			return startTime, http.StatusNotFound, nil, err
+		}
+
+		errResp := fmt.Errorf("Unable to retrieve %s: %s", tenmod.TenantMonitoredObjectKeysStr, err.Error())
+		return startTime, http.StatusInternalServerError, nil, errResp
+	}
+
+	responseMap := wrapJsonAPIObject(map[string]interface{}{"resourceIdentifiers": resourceIdentifierList}, "1", "filteredResourceIdentifierList")
+
+	responseBytes, err := json.Marshal(&responseMap)
+	if err != nil {
+		errResp := fmt.Errorf("Unable to convert %s data to json return format: %s", tenmod.TenantMonitoredObjectKeysStr, err.Error())
+		return startTime, http.StatusInternalServerError, nil, errResp
+	}
+
+	converted := swagmodels.MonitoredObjectFilteredListResponse{}
+	err = json.Unmarshal(responseBytes, &converted)
+	if err != nil {
+		errResp := fmt.Errorf("Unable to convert %s data to json return format: %s", tenmod.TenantMonitoredObjectKeysStr, err.Error())
+		return startTime, http.StatusInternalServerError, nil, errResp
+	}
+
+	reportAPICompletionState(startTime, http.StatusOK, mon.GetAllMonObjStr, mon.APICompleted, mon.TenantAPICompleted)
+	logger.Log.Infof("Retrieved %d %ss", len(converted.Data.Attributes.ResourceIdentifiers), tenmod.TenantMonitoredObjectKeysStr)
 	return startTime, http.StatusOK, &converted, nil
 }